Made some significant changes in order to fix a bug. Previously, Region::m_population was an unsigned long int. This resulted in bugs when populations grew to be large, as they charactaristically do. So this commit updates m_population to be an unsigned long long int, granting 64 bits to represent the population rather than 32. Changes in multiple classes were necessary to successfully implement this change. Utils::convertStringToUnsignedInt was updated to return an unsigned long long int, and the comparison between tmp and UINT32_MAX was updated to be a comparison betwen tmp and UINT64_MAX. There may still be bugs in computeTotalPopulation. These will be fixed in future commits.

+++ GeoRegions/Region.h
@@ -18,7 +18,7 @@ protected:
     unsigned int m_id = 0;
     RegionType m_regionType = UnknownRegionType;
     std::string m_name;
-    unsigned int m_population = 0;
+    unsigned long long int m_population = 0;
     double m_area = 0;
     bool m_isValid = false;
